Added cursor and sorting option to ICOs query

+++ src/services/graphql/schema/ico.js
@@ -12,7 +12,7 @@ export default gql`
   }
 
   extend type Query {
-    ICOs(cursor: Int, limit: Int): [ICO!]!
+    ICOs(cursor: Long, limit: Int): [ICO!]!
     ICO(_id: String!): ICO!
     averageValue(currency: String, exchange: String): Float!
     highestValue(currency: String, exchange: String): [ICO!]!
